reuse link carousel component
