+++ react-client/src/components/SimplifiedBroadcastUI/YoutubeVideoModal.js
@@ -3,433 +3,383 @@ import React from 'react'
 
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
-import { urlMetaData } from '../../redux/actions/convos.actions'
+import { RingLoader } from 'halogenium'
+import AddButton from './AddButton'
+import { downloadYouTubeVideo, uploadTemplate, urlMetaData } from '../../redux/actions/convos.actions'
+import { getVideoId } from '../../utility/utils'
+import YouTube from 'react-youtube'
 
 class YoutubeVideoModal extends React.Component {
-    constructor(props) {
-        super(props)
-        this.defaultErrorMsg = 'Please enter a valid YouTube video link'
-        this.elementLimit = 10
-        this.buttonLimit = 3
-        let cards = []
-        for (let i = 0; i < this.elementLimit; i++) {
-            if (props.cards && props.cards[i]) {
-                cards.push({ component: props.cards[i], disabled: false, id: i + 1 })
-            } else {
-                if (i === 0) {
-                    cards.push({
-                        disabled: true,
-                        id: i + 1,
-                        component: {
-                            title: '',
-                            subtitle: '',
-                            buttons: []
-                        }
-                    })
-                }
-            }
-        }
-        this.cardComponents = new Array(10)
-        this.state = {
-            cards,
-            links: props.links ? props.links : [{ valid: false, url: '', loading: false, errorMsg: this.defaultErrorMsg }],
-            selectedIndex: 0,
-            currentCollapsed: false,
-            disabled: props.edit ? false : true,
-            buttonActions: this.props.buttonActions ? this.props.buttonActions : ['open website', 'open webview'],
-            buttonDisabled: false,
-            actionDisabled: false,
-            numOfElements: cards.length
-        }
-
-        this.carouselIndicatorStyle = {
-            textIndent: '0',
-            margin: '0 2px',
-            width: '20px',
-            height: '20px',
-            border: 'none',
-            borderRadius: '100%',
-            lineHeight: '20px',
-            color: '#fff',
-            backgroundColor: '#999',
-            transition: 'all 0.25s ease'
-        }
-        this.carouselIndicatorActiveStyle = {
-            width: '25px',
-            height: '25px',
-            lineHeight: '25px',
-            backgroundColor: '#337ab7'
-        }
-        console.log('YoutubeVideoModal state in constructor', this.state)
-        console.log('YoutubeVideoModal props in constructor', this.props)
-        this.closeModal = this.closeModal.bind(this)
-        this.toggleHover = this.toggleHover.bind(this)
-        this.updateSelectedIndex = this.updateSelectedIndex.bind(this)
-        this.scrollToTop = this.scrollToTop.bind(this)
-        this.handleLinkChange = this.handleLinkChange.bind(this)
-        this.handleUrlMetaData = this.handleUrlMetaData.bind(this)
-        this.addLink = this.addLink.bind(this)
-        this.removeLink = this.removeLink.bind(this)
-        this.valid = this.valid.bind(this)
+  constructor(props) {
+    super(props)
+    console.log('YoutubeVideoModal props in constructor', props)
+    this.state = {
+      buttons: props.buttons.map(button => { return { visible: true, title: button.title } }),
+      numOfCurrentButtons: 0,
+      disabled: props.edit ? false : true,
+      buttonDisabled: false,
+      buttonLimit: 3,
+      buttonActions: this.props.buttonActions ? this.props.buttonActions : ['open website', 'open webview'],
+      file: this.props.file ? this.props.file : '',
+      link: this.props.youtubeLink ? this.props.youtubeLink : '',
+      videoLink: this.props.videoLink ? this.props.videoLink : '',
+      videoTitle: this.props.videoTitle ? this.props.videoTitle : '',
+      videoDescription: this.props.videoDescription ? this.props.videoDescription : '',
+      loading: false,
+      videoId: this.props.videoId ? this.props.videoId : null,
     }
+    console.log('YoutubeVideoModal state', this.state)
+    console.log('YoutubeVideoModal module', this.props.module)
+    this.updateFile = this.updateFile.bind(this)
+    this.handleDone = this.handleDone.bind(this)
+    this.updateButtonStatus = this.updateButtonStatus.bind(this)
+    this.validateYoutubeUrl = this.validateYoutubeUrl.bind(this)
+    this.handleLinkChange = this.handleLinkChange.bind(this)
+    this.uploadTemplate = this.uploadTemplate.bind(this)
+    this.setLoading = this.setLoading.bind(this)
+    this.closeModal = this.closeModal.bind(this)
+    this.handleUrlMetaData = this.handleUrlMetaData.bind(this)
+  }
 
-    valid() {
-        let links = this.state.links
-        for (let i = 0; i < links.length; i++) {
-            if (!links[i].valid || links[i].loading) {
-                return false
-            }
-        }
-        return true
+  handleLinkChange(e) {
+    console.log('changing link', e.target.value)
+    this.setState({ buttons: [], disabled: true, link: e.target.value, videoId: null, videoTitle: null, videoDescription: null }, () => {
+      this.validateYoutubeUrl()
+    })
+  }
+
+  handleDone() {
+    if (this.props.noButtons) {
+      this.addComponent([])
+    } else {
+      this.AddButton.handleDone()
     }
+  }
 
-    removeLink(index) {
-        let links = this.state.links
-        let cards = this.state.cards
-        let selectedIndex = this.state.selectedIndex
-        if (links.length > 1) {
-            links.splice(index, 1)
-            if (cards[index]) {
-                cards.splice(index, 1)
-            }
-            if (selectedIndex === 0) {
-                selectedIndex = 1
-            } else {
-                selectedIndex = index - 1
-            }
-            this.setState({ links, cards, selectedIndex })
-        }
+  addComponent(buttons) {
+    console.log('addComponent YoutubeVideoModal', this.state)
+    if (this.props.module === 'whatsapp') {
+      this.props.addComponent({
+        id: this.props.id >= 0 ? this.props.id : null,
+        componentType: 'text',
+        videoTitle: this.state.videoTitle,
+        videoDescription: this.state.videoDescription,
+        videoId: this.state.videoId,
+        text: this.state.link,
+        buttons
+      }, this.props.edit)
+    } else {
+      this.props.addComponent({
+        id: this.props.id,
+        youtubeLink: this.state.link,
+        videoLink: this.state.videoLink,
+        componentType: 'media',
+        fileurl: this.state.file.fileurl,
+        fileName: this.state.file.fileName,
+        image_url: '',
+        size: this.state.file.size,
+        type: this.state.file.type,
+        mediaType: 'video',
+        buttons
+      })
     }
+  }
 
+  updateButtonStatus(status) {
+    console.log('updateButtonStatus MediaModal', status)
+    this.setState(status)
+  }
 
-    addLink() {
-        let links = this.state.links
-        if (links.length < this.elementLimit) {
-            links.push({ url: '', valid: false, loading: false, errorMsg: this.defaultErrorMsg })
-            this.setState({ links })
-        }
+  updateFile(file) {
+    if (file === 'ERR_LIMIT_REACHED') {
+      this.setState({ fileSizeExceeded: true, disabled: true, loading: false })
+      return
     }
-
-    toggleHover(index, hover) {
-        this.setState({ hover: hover ? index : -1 })
+    console.log('updating YouTube file', file)
+    if (this.props.module !== 'jsonads') {
+      this.uploadTemplate(file)
+    } else {
+      this.setState({
+        loading: false,
+        file: file,
+        videoLink: file.fileurl.url
+      }, () => {
+        this.refs.video.pause();
+        this.refs.video.load();
+        this.refs.video.play();
+      })
     }
+    // this.setState({file, videoLink: file.fileurl.url}, () => {
+    //   this.refs.video.pause();
+    //   this.refs.video.load();
+    //   this.refs.video.play();
+    // })
+  }
 
-    componentDidMount() {
-        //Improve Later
-        let that = this
-        $('#carouselExampleControls').on('slide.bs.carousel', function (e) {
-            var active = $(e.target).find('.carousel-inner > .carousel-item.active');
-            var from = active.index();
-            var next = $(e.relatedTarget);
-            var to = next.index();
-            console.log(from + ' => ' + to);
-            that.setState({ selectedIndex: to })
+  setLoading() {
+    this.setState({ loading: true })
+  }
+
+  uploadTemplate(data) {
+    console.log('updating file AddMedia', data)
+    //this.props.updateFile(data)
+    var video = data.type.match('video.*')
+    if (video) {
+      console.log('video uploading template')
+      this.props.uploadTemplate({
+        pages: this.props.pages,
+        url: data.fileurl.url,
+        componentType: 'video',
+        id: data.fileurl.id,
+        name: data.fileurl.name
+      }, {
+        id: this.props.id,
+        componentType: 'video',
+        fileName: data.fileurl.name,
+        type: data.fileurl.type,
+        size: data.fileurl.size
+      }, (newData) => this.updateFileUrl(data, newData))
+    }
+  }
+
+  updateFileUrl(data, newData) {
+    data.fileurl = newData.fileurl
+    console.log('updating fileurl of YoutubeVideoModal', data)
+    this.setState({
+      loading: false,
+      file: data,
+      videoLink: data.fileurl.url
+    }, () => {
+      this.refs.video.pause();
+      this.refs.video.load();
+      this.refs.video.play();
+    })
+  }
+
+
+  validateYoutubeUrl() {
+    let url = this.state.link
+    if (url !== undefined || url !== '') {
+      /* eslint-disable */
+      var regExp = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|\&v=|\?v=)([^#\&\?]*).*/
+      /* eslint-enable */
+      var match = url.match(regExp)
+      if (match && match[2].length === 11) {
+        this.setState({ disabled: false, loading: true, fileSizeExceeded: false, edited: true }, () => {
+          if (this.props.module && this.props.module === 'whatsapp') {
+            console.log('setting videoId')
+            let videoId = getVideoId(this.state.link)
+            console.log('setting videoId', videoId)
+            this.setState({ videoId, loading: false })
+            this.props.urlMetaData(this.state.link, (data) => this.handleUrlMetaData(data))
+          } else {
+            this.props.downloadYouTubeVideo(this.state.link, this.props.id, (file) => { this.updateFile(file) })
+          }
         })
+      } else {
+        this.setState({ disabled: true, file: null, fileSizeExceeded: false, edited: false })
+      }
     }
+  }
 
-    addComponent() {
-        console.log('addComponent this.state.cards', this.state.cards)
-        if (this.state.cards.length === 1) {
-            let card = this.state.cards[0].component
-            this.props.addComponent({
-                id: this.props.id,
-                links: this.state.links,
-                youtubeVideo: true,
-                componentType: 'card',
-                image_url: card.image_url ? card.image_url : '',
-                title: card.title,
-                description: card.subtitle,
-                buttons: card.buttons
-            }, this.props.edit)
-        } else if (this.state.cards.length > 1) {
-            let cards = this.state.cards.map((card, index) => {
-                return {
-                    id: card.id ? card.id : '',
-                    image_url: card.component.image_url ? card.component.image_url : '',
-                    title: card.component.title,
-                    subtitle: card.component.subtitle ? card.component.subtitle : card.component.description,
-                    buttons: card.component.buttons
-                }
-            })
-            this.props.addComponent({
-                id: this.props.id,
-                links: this.state.links,
-                componentType: 'gallery',
-                cards
-            }, this.props.edit)
-        }
+  handleUrlMetaData(data) {
+    this.setState({ videoTitle: data.title, videoDescription: data.description })
+  }
+
+  closeModal() {
+    if (!this.state.edited) {
+      this.props.closeModal()
+    } else {
+      this.props.showCloseModalAlertDialog()
     }
+  }
 
 
-    closeModal () {
-        if (!this.state.edited) {
-            this.props.closeModal()
-        } else {
-            this.props.showCloseModalAlertDialog()
-        }
-    }
+  UNSAFE_componentWillUnmount() {
+    this.props.closeModal()
+  }
 
-    scrollToTop(elementId) {
-        document.getElementById(elementId).scrollIntoView({ behavior: 'smooth' })
-    }
+  // UNSAFE_componentWillReceiveProps(nextProps) {
+  //   console.log('componentWillRecieveProps YoutubeVideoModal')
+  //   this.setState({
+  //     buttons: nextProps.buttons.map(button => { return { visible: true, title: button.title } }),
+  //     numOfCurrentButtons: 0,
+  //     disabled: nextProps.edit ? false : true,
+  //     buttonDisabled: false,
+  //     buttonLimit: 3,
+  //     buttonActions: nextProps.buttonActions ? nextProps.buttonActions : ['open website', 'open webview'],
+  //     file: nextProps.file ? this.props.file : '',
+  //     link: nextProps.youtubeLink ? nextProps.youtubeLink : '',
+  //     videoLink: nextProps.videoLink ? nextProps.videoLink : '',
+  //     videoTitle: nextProps.videoTitle ? nextProps.videoTitle : '',
+  //     videoDescription: nextProps.videoDescription ? nextProps.videoDescription : '',
+  //     loading: false,
+  //     videoId: nextProps.videoId ? nextProps.videoId : null,
+  //   })
+  // }
 
-    updateSelectedIndex(index) {
-        this.setState({ selectedIndex: index })
-    }
-
-    handleUrlMetaData(data, index) {
-        console.log('url meta data retrieved', data)
-        let links = this.state.links
-        let cards = this.state.cards
-        if (!data || !data.ogTitle || !data.ogDescription || !data.ogImage || !data.ogImage.url) {
-            let errorMsg = ''
-            if (!data) {
-                errorMsg = 'Invalid YouTube link'
-            } else if (data.ogImage && !data.ogImage.url) { 
-                errorMsg = 'No default image found'
-            } else {
-                errorMsg = 'Not enough metadata present in link'
-            }
-            links[index] = Object.assign(links[index], { loading: false, valid: false, errorMsg })
-            cards[index] = {
-                disabled: true,
-                id: index + 1,
-                component: {
-                    title: '',
-                    subtitle: '',
-                    buttons: []
-                }
-            }
-            this.setState({ links, cards })
-        } else {
-            if (data.ogImage.url.startsWith('/')) {
-                data.ogImage.url = links[index].url + data.ogImage.url
-            }
-            cards[index] = {
-                id: index + 1,
-                component: {
-                    title: data.ogTitle.length > 80 ? data.ogTitle.substring(0, 80) + '...' : data.ogTitle,
-                    subtitle: data.ogDescription.length > 80 ? data.ogDescription.substring(0, 80) + '...' : data.ogDescription,
-                    image_url: data.ogImage.url,
-                    buttons: this.props.hideWebUrl ? [] :[
-                        {
-                            title: 'Watch on YouTube',
-                            type: 'web_url',
-                            url: links[index].url
-                        }
-                    ]
-                }
-            }
-            links[index] = Object.assign(links[index], { loading: false, valid: true, errorMsg: this.defaultErrorMsg })
-            this.setState({ links, cards, selectedIndex: index })
-        }
-    }
-
-    handleLinkChange(e, index) {
-        console.log('changing link', e.target.value)
-        let link = e.target.value
-        let links = this.state.links
-        let cards = this.state.cards
-        if (this.state.links.length < this.elementLimit) {
-            if (this.validateYoutubeURL(link)) {
-                links[index] = { url: link, valid: true, loading: true, errorMsg: this.defaultErrorMsg }
-            } else {
-                cards[index] = {
-                    disabled: true,
-                    id: index + 1,
-                    component: {
-                        title: '',
-                        subtitle: '',
-                        buttons: []
-                    }
-                }
-                links[index] = { url: link, valid: false, loading: false, errorMsg: this.defaultErrorMsg }
-            }
-        }
-        this.setState({ links, cards, edited: true }, () => {
-            if (links[index].valid) {
-                this.props.urlMetaData(link, (data) => this.handleUrlMetaData(data, index))
-            }
-        })
-    }
-
-    validateYoutubeURL (url) {
-        if (url !== undefined || url !== '') {
-            var regExp = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|\&v=|\?v=)([^#\&\?]*).*/
-            var match = url.match(regExp)
-            return match && match[2].length === 11
-        } else {
-            return false
-        }
-    }
-
-    UNSAFE_componentWillUnmount() {
-        this.props.closeModal()
-    }
-
-    render () {
-        return (
-            <div className="modal-content" style={{ width: '72vw' }}>
-                <div style={{ display: 'block' }} className="modal-header">
-                    <h5 className="modal-title" id="exampleModalLabel">
-                        Youtube Video
-                    </h5>
-                    <button style={{ marginTop: '-10px', opacity: '0.5', color: 'black' }} type="button" className="close" onClick={this.closeModal} aria-label="Close">
-                        <span aria-hidden="true">
-                            &times;
-                        </span>
-                    </button>
-                </div>
-                <div style={{ color: 'black' }} className="modal-body">
-                  <div className='row'>
-                    <div className='col-6' style={{maxHeight: '65vh', overflowY: 'scroll'}}>
-                        {
-                            this.state.links.map((link, index) => (
-                                    <div>
-                                        <div className='row'>
-                                            <div className='col-11'>
-                                                <input value={link.url} style={{ maxWidth: '100%', borderColor: !link.valid && !this.state.loading ? 'red' : (this.state.loading || link.valid) ? 'green' : ''}} onChange={(e) => this.handleLinkChange(e, index)} className='form-control' />
-                                            </div>
-
-                                            <div className='col-1'>
-                                                <div onClick={() => this.removeLink(index)} style={{marginTop: '10px', cursor: 'pointer'}}><span role='img' aria-label='times'>❌</span></div>
-                                            </div>
-
-                                        </div>
-                                        <div style={{color: 'green'}}>{link.valid && !link.loading ? '*Youtube video link is valid.' : ''}</div>
-                                        <div style={{color: 'red'}}>{!link.valid && !link.loading ? `*${link.errorMsg}` : ''}</div>
-                                        <div style={{marginBottom: '30px', color: 'green'}}>{link.loading ? '*Retrieving youtube video meta data.' : ''}</div>
-                                    </div>
-                                ))
-                        }
-                        {/* {
-                            this.state.links.length < this.elementLimit &&
-                            <div className='ui-block hoverborder'
-                                style={{minHeight: '30px',
-                                width: '100%',
-                                marginLeft: '0px',
-                                borderColor: this.props.required && visibleButtons.length === 0 ? 'red' : ''}}
-                                onClick={this.addLink}>
-                                <div style={{paddingTop: '5px'}} className='align-center'>
-                                <h6> + Add Link </h6>
-                            </div>
-                        </div>
-                        } */}
-                    </div>
-                    <div className='col-1'>
-                    <div style={{minHeight: '100%', width: '1px', borderLeft: '1px solid rgba(0,0,0,.1)'}} />
-                    </div>
-                    <div className='col-5'>
-                    <h4 style={{marginLeft: '-50px'}}>Preview:</h4>
-                    <div className='ui-block' style={{overflowY: 'auto', border: '1px solid rgba(0,0,0,.1)', borderRadius: '3px', maxHeight: '68vh', minHeight: '68vh', marginLeft: '-50px'}} >
-                      <div id="carouselExampleControls" data-interval="false" className="carousel slide ui-block" data-ride="carousel">
-
-                        {
-                            this.state.cards.length > 1 &&
-                            <ol className="carousel-indicators carousel-indicators-numbers" style={{bottom: '-65px'}}>
-                                {
-                                this.state.cards.map((card, index) => {
-                                    return (<li
-                                        style={(this.state.hover === index || this.state.selectedIndex === index) ? {...this.carouselIndicatorStyle, ...this.carouselIndicatorActiveStyle} : this.carouselIndicatorStyle}
-                                        onMouseEnter={() => this.toggleHover(index, true)}
-                                        onMouseLeave={() => this.toggleHover(index, false)}
-                                        data-target="#carouselExampleControls"
-                                        data-slide-to={index}
-                                        onClick={() => this.updateSelectedIndex(index)}
-                                        className={(index === this.state.selectedIndex ? "active" : "")}>
-                                        {index+1}
-                                    </li>)
-                                })
-                                }
-                            </ol>
-                        }
-                        <div className="carousel-inner">
-                        {
-                            this.state.cards.map((card, index) => {
-                            return (
-                                <div style={{border: '1px solid rgba(0,0,0,.1)', borderRadius: '10px', minHeight: '200px', maxWidth: '250px', margin: 'auto', marginTop: '60px'}} className={"carousel-item " + (index === this.state.selectedIndex ? "active" : "") + (index === this.state.selectedIndex+1 ? "next" : "") + (index === this.state.selectedIndex-1 ? "prev" : "")}>
-                                    {
-                                        card.component.image_url &&
-                                        <img alt='' src={card.component.image_url} style={{maxHeight: '140px', minWidth: '250px', padding: '20px', paddingTop: '30px', margin: '-25px'}} />
-                                    }
-                                    <hr style={{marginTop: card.component.image_url ? '' : '100px', marginBottom: '5px'}} />
-                                    <h6 style={{textAlign: 'left', marginLeft: '10px', marginTop: '10px', fontSize: '16px'}}>{card.component.title}</h6>
-                                    <p style={{textAlign: 'left', marginLeft: '10px', marginTop: '5px', fontSize: '13px'}}>{card.component.subtitle ? card.component.subtitle : card.component.description}</p>
-                                    <p style={{textAlign: 'left', marginLeft: '10px', fontSize: '13px'}}>{card.component.default_action && card.component.default_action.url}</p>
-                                    {
-                                        card.component.buttons && card.component.buttons.map((button, index) => (
-                                        (button.visible || button.type) && (
-                                          <div>
-                                              <hr style={{marginTop: !card.component.title && !card.component.subtitle && index === 0 ? '50px' : ''}}/>
-                                              <h5 style={{color: '#0782FF'}}>{button.title}</h5>
-                                          </div>
-                                        )
-                                      ))
-                                    }
-                                </div>
-                            )
-                            })
-                        }
-                        </div>
-                        {
-                            this.state.cards.length > 1 &&
-                            <div>
-                            {
-                              this.state.selectedIndex > 0 &&
-                              <a href='#/' onClick={(e) => this.updateSelectedIndex(this.state.selectedIndex-1)} className="carousel-control-prev" style={{top: '125px'}} role="button">
-                               <span className="carousel-control-prev-icon" style={{cursor: 'pointer', backgroundImage: `url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23000' viewBox='0 0 8 8'%3E%3Cpath d='M5.25 0l-4 4 4 4 1.5-1.5-2.5-2.5 2.5-2.5-1.5-1.5z'/%3E%3C/svg%3E")`}} aria-hidden="true"></span>
-                               <span className="sr-only">Previous</span>
-                              </a>
-                            }
-                            {
-                              this.state.selectedIndex < this.state.cards.length-1 &&
-                              <a href='#/' onClick={(e) => this.updateSelectedIndex(this.state.selectedIndex+1)} className="carousel-control-next" style={{top: '125px'}} role="button" >
-                                <span className="carousel-control-next-icon" style={{cursor: 'pointer', backgroundImage: `url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23000' viewBox='0 0 8 8'%3E%3Cpath d='M2.75 0l-1.5 1.5 2.5 2.5-2.5 2.5 1.5 1.5 4-4-4-4z'/%3E%3C/svg%3E")`}} aria-hidden="true"></span>
-                                <span className="sr-only">Next</span>
-                              </a>
-
-                            }
-                            </div>
-                        }
-                      </div>
-                    </div>
-                    </div>
-                    <div className='row' style={{marginTop: '-5vh'}}>
-                    { this.props.module !== 'commentcapture'
-                        ? <div className='pull-right'>
-                            <button onClick={this.closeModal} className='btn btn-primary' style={{marginRight: '25px', marginLeft: '280px'}}>
-                                Cancel
-                            </button>
-                            <button disabled={!this.valid()} onClick={() => this.addComponent()} className='btn btn-primary'>
-                                {this.props.edit ? 'Edit' : 'Next'}
-                            </button>
-                        </div>
-                        : <div className='pull-right'>
-                            <button onClick={this.closeModal} className='btn btn-secondary' style={{marginRight: '25px', marginLeft: '280px'}}>
-                                Cancel
-                            </button>
-                            <button disabled={!this.valid()} onClick={() => {this.props.saveLinks(this.state.links,this.state.cards)}} className='btn btn-primary'>
-                                Save
-                            </button>
-                        </div>
-                    }
-
-                    </div>
-                  </div>
-                </div>
+  render() {
+    console.log('video link', this.state.link)
+    console.log('YoutubeVideoModal props in render', this.props)
+    let visibleButtons = this.state.buttons.filter(button => button.visible)
+    return (
+      <div className="modal-content" style={{width: '72vw'}}>
+        <div style={{ display: 'block' }} className="modal-header">
+          <h5 className="modal-title" id="exampleModalLabel">
+            Add Video from YouTube
+									</h5>
+          <button style={{ marginTop: '-10px', opacity: '0.5', color: 'black' }} type="button" className="close" onClick={this.closeModal} aria-label="Close">
+            <span aria-hidden="true">
+              &times;
+											</span>
+          </button>
+        </div>
+        <div style={{ color: 'black' }} className="modal-body">
+          <div className='row'>
+            <div className='col-6' style={{ maxHeight: '65vh', overflowY: 'scroll' }}>
+              <h4>YouTube Link:</h4>
+              <input value={this.state.link} style={{ maxWidth: '100%', borderColor: this.state.disabled && !this.state.loading ? 'red' : (this.state.loading || !this.state.disabled) ? 'green' : '' }} onChange={this.handleLinkChange} className='form-control' />
+              <div style={{ color: 'red' }}>{this.state.fileSizeExceeded ? '*The size of this YouTube video exceeds the 25 Mb limit imposed by Facebook. Please try another video.' : ''}</div>
+              <div style={{ color: 'red' }}>{!this.state.fileSizeExceeded && this.state.disabled && !this.state.loading ? '*Please enter a valid YouTube link.' : ''}</div>
+              <div style={{ marginBottom: '30px', color: 'green' }}>{this.state.loading ? '*Please wait for the YouTube video to download.' : ''}</div>
+              {
+                this.state.file && this.props.module !== 'whatsapp' &&
+                <AddButton
+                  replyWithMessage={this.props.replyWithMessage}
+                  disabled={this.state.disabled}
+                  buttons={this.state.buttons}
+                  finalButtons={this.props.buttons}
+                  buttonLimit={this.state.buttonLimit}
+                  pageId={this.props.pageId}
+                  buttonActions={this.state.buttonActions}
+                  ref={(ref) => { this.AddButton = ref }}
+                  updateButtonStatus={this.updateButtonStatus}
+                  toggleGSModal={this.props.toggleGSModal}
+                  closeGSModal={this.props.closeGSModal}
+                  addComponent={(buttons) => this.addComponent(buttons)} />
+              }
             </div>
-        )
-    }
+            <div className='col-1'>
+              <div style={{ minHeight: '100%', width: '1px', borderLeft: '1px solid rgba(0,0,0,.1)' }} />
+            </div>
+            <div className='col-5'>
+              <h4 style={{ marginLeft: '-50px' }}>Preview:</h4>
+              <div className='ui-block' style={{ overflowY: 'auto', border: '1px solid rgba(0,0,0,.1)', borderRadius: '3px', minHeight: '68vh', maxHeight: '68vh', marginLeft: '-50px' }} >
+                <div className='ui-block' style={{ border: !this.state.disabled ? '1px solid rgba(0,0,0,.1)' : '', borderRadius: '10px', maxWidth: '80%', margin: 'auto', marginTop: '80px' }} >
+                  <div style={{ border: '1px solid rgba(0,0,0,.1)', borderRadius: '10px', minHeight: '200px', maxWidth: '250px', margin: 'auto', marginTop: '60px' }} className={"carousel-item " + (index === this.state.selectedIndex ? "active" : "") + (index === this.state.selectedIndex + 1 ? "next" : "") + (index === this.state.selectedIndex - 1 ? "prev" : "")}>
+                      {
+                        card.component.image_url &&
+                        <img alt='' src={card.component.image_url} style={{objectFit: 'cover', minHeight: '170px', maxHeight: '170px', maxWidth: '300px', paddingBottom: '11px', paddingTop: '29px', margin: '-25px', width: '100%', height: '100%' }} />
+                      }
+                      <hr style={{ marginTop: card.component.image_url ? '' : '100px', marginBottom: '5px' }} />
+                      <h6 style={{ textAlign: 'justify', marginLeft: '10px', marginTop: '10px', fontSize: '16px' }}>{card.component.title}</h6>
+                      <p style={{ textAlign: 'justify', marginLeft: '10px', marginTop: '5px', fontSize: '13px' }}>{card.component.subtitle ? card.component.subtitle : card.component.description}</p>
+                      <p style={{ textAlign: 'justify', marginLeft: '10px', fontSize: '13px' }}>{card.component.default_action && card.component.default_action.url}</p>
+                      {
+                        card.component.buttons.map((button, index) => (
+                          (button.visible || button.type) && (
+                              <div>
+                                <hr style={{ marginTop: !card.component.title && !card.component.subtitle && index === 0 ? '50px' : '' }} />
+                                <h5 style={{ color: '#0782FF' }}>{button.title}</h5>
+                              </div>
+                            )
+                        ))
+                      }
+                    </div>
+                  
+                  
+                  {
+                    this.state.loading && <div className='align-center' style={{ padding: '50px' }}>
+                      <center><RingLoader color='#FF5E3A' /></center>
+                    </div>
+                  }
+
+                  {
+                    this.state.videoId &&
+                    <div>
+                      <div>
+                        <YouTube
+                          videoId={this.state.videoId}
+                          opts={{
+                            height: '150',
+                            width: '300',
+                            playerVars: {
+                              autoplay: 0
+                            }
+                          }}
+                        />
+                      </div>
+                      {
+                        this.state.videoTitle &&
+                        <div>
+                          <div style={{ textAlign: 'left', 'marginLeft': '10px' }}>
+                            <h6>{this.state.videoTitle}</h6>
+                            <h7>{this.state.videoDescription.length > 50 ? this.state.videoDescription.substr(0, 47) + '...' : this.state.videoDescription}</h7>
+                            <p style={{ fontSize: '0.7em', marginTop: '5px' }}>www.youtube.com</p>
+                          </div>
+                          <hr />
+                          <p style={{ textAlign: 'left', color: '#0782FF', fontSize: '0.9em', marginLeft: '5px' }}>{this.state.link}</p>
+                        </div>
+                      }
+                    </div>
+                  }
+                  {
+                    (this.props.module !== 'whatsapp' && !this.state.disabled && !this.state.loading) &&
+                    <video ref="video" controls style={{ width: '100%', borderRadius: '10px', marginTop: '-10px', borderBottomLeftRadius: '0px', borderBottomRightRadius: '0px' }} name='media' id='youtube_player'>
+                      <source src={this.state.videoLink} type='audio/mpeg' />
+                    </video>
+                  }
+                  {
+                    visibleButtons.map((button, index) => {
+                      let style = null
+                      if (index === 0 || visibleButtons.length === 1) {
+                        style = { marginTop: '55%' }
+                        if (this.state.link && !this.state.disabled) {
+                          style = { marginTop: '-9%' }
+                        }
+                      }
+                      return (
+                        <div style={style}>
+                          <hr />
+                          <h5 style={{ color: '#0782FF' }}>{button.title}</h5>
+                        </div>
+                      )
+                    })
+                  }
+
+                </div>
+              </div>
+            </div>
+            <div className='col-6' style={{ marginTop: '-5vh' }}>
+              <div className='pull-right'>
+              <button onClick={this.closeModal} className='btn btn-primary' style={{ marginRight: '20px' }}
+              >
+                  Cancel
+                </button>
+                <button disabled={(!this.props.module === 'whatsapp' && !this.state.file) || this.state.disabled || this.state.buttonDisabled || (this.props.module === 'whatsapp' && !this.state.videoTitle) || this.state.loading} onClick={() => this.handleDone()} className='btn btn-primary'
+                 >
+                  {this.props.edit ? 'Edit' : 'Next'}
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    )
+  }
 }
 
 function mapStateToProps(state) {
-    return {
-    }
+  console.log(state)
+  return {
+  }
 }
 
 function mapDispatchToProps(dispatch) {
-    return bindActionCreators({
-        urlMetaData
-    }, dispatch)
+  return bindActionCreators({
+    uploadTemplate,
+    downloadYouTubeVideo,
+    urlMetaData
+  }, dispatch)
 }
 export default connect(mapStateToProps, mapDispatchToProps)(YoutubeVideoModal)
